+++ spec/helper/TestsHelper.ts
@@ -1,5 +1,5 @@
-import { BrowserUtils} from '../../src/ts/BrowserUtils';
-jest.mock('../../src/ts/BrowserUtils');
+import { BrowserUtils} from '../../src/ts/helper/BrowserUtils';
+jest.mock('../../src/ts/helper/BrowserUtils');
 
 beforeAll(() => {
   BrowserUtils.isMobile = jest.fn().mockReturnValue(false);